proce number formate update

+++ database/migrations/2025_04_23_085132_create_plans_table.php
@@ -13,8 +13,7 @@ return new class extends Migration {
             $table->id();
 
             $table->enum('subscription_plan', ['free', 'monthly', 'yearly', 'lifetime']);
-            // $table->decimal('price', 8, 2);
-            $table->float('price', 8);
+            $table->decimal('price', 8, 2);
             $table->enum('billing_cycle', ['monthly', 'yearly', 'lifetime']);
             $table->integer('touch_points')->nullable()->default(null);
             $table->boolean('has_ads');
